refactor(xpra): remove raw packet handling quirk
In the original codebase the raw packet queue was a hybrid between an
object and array in order to get the length of a sparse list.

This has been changed to be an array exclusively with a filter because
the performance overhead is quite low and makes it easier to understand
the code.